refactor: code quality
* extract method
* use base class for game baords
